223: Add admin boolean

+++ db/migrate/20151218195706_drop_genre_table.rb
@@ -1,5 +1,5 @@
 class DropGenreTable < ActiveRecord::Migration
   def change
-    drop_table :genres
+    drop_table :genres if table_exists?(:genres)
   end
 end
